requirements.txt
streamlit
pandas
matplotlib
python-dotenv
openai
spotipy